chore: removes cyclic dependency between core/config and core/create
